refactor(OIRecognizer) always use send function

+++ src/recognizer/OIRecognizer.ts
@@ -60,8 +60,8 @@ export class OIRecognizer
   protected currentPartId?: string
   protected currentErrorCode?: string | number
 
-  protected connected: DeferredPromise<void>
-  protected initialized: DeferredPromise<void>
+  protected connected?: DeferredPromise<void>
+  protected initialized?: DeferredPromise<void>
 
   protected addStrokeDeferred?: DeferredPromise<TOIMessageEventGesture | undefined>
   protected contextlessGestureDeferred: Map<string, DeferredPromise<TOIMessageEventContextlessGesture>>
@@ -87,9 +87,6 @@ export class OIRecognizer
 
     this.exportDeferredMap = new Map()
     this.contextlessGestureDeferred = new Map()
-
-    this.connected = new DeferredPromise<void>()
-    this.initialized = new DeferredPromise<void>()
   }
 
   get mimeTypes(): string[]
@@ -102,11 +99,12 @@ export class OIRecognizer
     return InternalEvent.getInstance()
   }
 
-  #send(message: TOIMessageEvent): void
+  async #send(message: TOIMessageEvent): Promise<void>
   {
     if (!this.socket) {
       throw new Error("Recognizer must be initilized")
     }
+    await this.connected?.promise
     if (this.socket.readyState === this.socket.OPEN) {
       this.socket.send(JSON.stringify(message))
     }
@@ -117,7 +115,7 @@ export class OIRecognizer
 
   protected rejectDeferredPending(error: Error | string): void
   {
-    this.initialized.reject(error)
+    this.initialized?.reject(error)
     this.addStrokeDeferred?.reject(error)
     this.transformStrokeDeferred?.reject(error)
     this.eraseStrokeDeferred?.reject(error)
@@ -140,8 +138,8 @@ export class OIRecognizer
 
   protected resetAllDeferred(): void
   {
-    this.connected = new DeferredPromise<void>()
-    this.initialized = new DeferredPromise<void>()
+    this.connected = undefined
+    this.initialized = undefined
     this.addStrokeDeferred = undefined
     this.contextlessGestureDeferred.clear()
     this.transformStrokeDeferred = undefined
@@ -223,9 +221,9 @@ export class OIRecognizer
 
   protected openCallback(): void
   {
-    this.connected.resolve()
+    this.connected?.resolve()
     this.reconnectionCount = 0
-    this.#send({
+    this.send({
       type: "authenticate",
       "myscript-client-name": "iink-ts",
       "myscript-client-version": "1.0.0-buildVersion",
@@ -234,7 +232,7 @@ export class OIRecognizer
 
   protected async manageHMACChallenge(hmacChallengeMessage: TOIMessageEventHMACChallenge): Promise<void>
   {
-    this.#send({
+    this.send({
       type: "hmac",
       hmac: await computeHmac(hmacChallengeMessage.hmacChallenge, this.serverConfiguration.applicationKey, this.serverConfiguration.hmacKey)
     })
@@ -248,21 +246,23 @@ export class OIRecognizer
     })
     this.pingWorker.onmessage = () =>
     {
-      if (this.pingCount < this.serverConfiguration.websocket.maxPingLostCount) {
-        this.#send({ type: "ping" })
+      if (this.socket.readyState <= 1) {
+        if (this.pingCount < this.serverConfiguration.websocket.maxPingLostCount) {
+          this.send({ type: "ping" })
+        }
+        else {
+          this.close(1000, "MAXIMUM_PING_REACHED")
+          this.pingWorker?.terminate()
+        }
+        this.pingCount++
       }
-      else {
-        this.close(1000, "MAXIMUM_PING_REACHED")
-        this.pingWorker?.terminate()
-      }
-      this.pingCount++
     }
   }
 
   protected manageAuthenticated(): void
   {
     const pixelTomm = 25.4 / 96
-    this.#send({
+    this.send({
       type: this.sessionId ? "restoreSession" : "initSession",
       iinkSessionId: this.sessionId,
       scaleX: pixelTomm,
@@ -277,16 +277,16 @@ export class OIRecognizer
       this.sessionId = sessionDescriptionMessage.iinkSessionId
     }
     if (this.currentPartId) {
-      this.#send({ type: "openContentPart", id: this.currentPartId })
+      this.send({ type: "openContentPart", id: this.currentPartId })
     }
     else {
-      this.#send({ type: "newContentPart", contentType: this.recognitionConfiguration.type, mimeTypes: this.mimeTypes })
+      this.send({ type: "newContentPart", contentType: this.recognitionConfiguration.type, mimeTypes: this.mimeTypes })
     }
   }
 
   protected manageNewPartMessage(newPartMessage: TOIMessageEventNewPart): void
   {
-    this.initialized.resolve()
+    this.initialized?.resolve()
     this.currentPartId = newPartMessage.id
   }
 
@@ -429,12 +429,14 @@ export class OIRecognizer
       this.sessionId = undefined
       this.currentPartId = undefined
     }
+    this.connected = new DeferredPromise<void>()
+    this.initialized = new DeferredPromise<void>()
     this.socket = new WebSocket(this.url)
     this.clearSocketListener()
     this.socket.addEventListener("open", this.openCallback.bind(this))
     this.socket.addEventListener("close", this.closeCallback.bind(this))
     this.socket.addEventListener("message", this.messageCallback.bind(this))
-    await this.initialized.promise
+    await this.initialized?.promise
     if (this.serverConfiguration.websocket.pingEnabled) {
       this.pingCount = 0
       this.initPing()
@@ -450,7 +452,6 @@ export class OIRecognizer
     switch (this.socket.readyState) {
       case this.socket.CONNECTING:
       case this.socket.OPEN:
-        await this.initialized.promise
         this.#send(message)
         return Promise.resolve()
       case this.socket.CLOSING:
@@ -461,7 +462,7 @@ export class OIRecognizer
             this.event.emitClearMessage()
             await this.init()
             await this.waitForIdle()
-            return this.send(message)
+            return this.#send(message)
           }
           else {
             return Promise.reject(new Error("Unable to send message. The maximum number of connection attempts has been reached."))
@@ -489,6 +490,7 @@ export class OIRecognizer
       this.addStrokeDeferred.resolve(undefined)
       return this.addStrokeDeferred?.promise
     }
+    await this.initialized?.promise
     await this.send(this.buildAddStrokesMessage(strokes, processGestures))
     return this.addStrokeDeferred?.promise
   }
@@ -508,6 +510,7 @@ export class OIRecognizer
       this.replaceStrokeDeferred.resolve()
       return this.replaceStrokeDeferred?.promise
     }
+    await this.initialized?.promise
     await this.send(this.buildReplaceStrokesMessage(oldStrokeIds, newStrokes))
     return this.replaceStrokeDeferred?.promise
   }
@@ -529,6 +532,7 @@ export class OIRecognizer
       this.transformStrokeDeferred.resolve()
       return this.transformStrokeDeferred?.promise
     }
+    await this.initialized?.promise
     await this.send(this.buildTransformTranslateMessage(strokeIds, tx, ty))
     return this.transformStrokeDeferred?.promise
   }
@@ -551,6 +555,7 @@ export class OIRecognizer
       this.transformStrokeDeferred.resolve()
       return this.transformStrokeDeferred?.promise
     }
+    await this.initialized?.promise
     await this.send(this.buildTransformRotateMessage(strokeIds, angle, x0, y0))
     return this.transformStrokeDeferred?.promise
   }
@@ -574,6 +579,7 @@ export class OIRecognizer
       this.transformStrokeDeferred.resolve()
       return this.transformStrokeDeferred?.promise
     }
+    await this.initialized?.promise
     await this.send(this.buildTransformScaleMessage(strokeIds, scaleX, scaleY, x0, y0))
     return this.transformStrokeDeferred?.promise
   }
@@ -594,6 +600,7 @@ export class OIRecognizer
       this.transformStrokeDeferred.resolve()
       return this.transformStrokeDeferred?.promise
     }
+    await this.initialized?.promise
     await this.send(this.buildTransformMatrixMessage(strokeIds, matrix))
     return this.transformStrokeDeferred?.promise
   }
@@ -612,6 +619,7 @@ export class OIRecognizer
       this.eraseStrokeDeferred.resolve()
       return this.eraseStrokeDeferred?.promise
     }
+    await this.initialized?.promise
     await this.send(this.buildEraseStrokesMessage(strokeIds))
     return this.eraseStrokeDeferred?.promise
   }
@@ -623,6 +631,7 @@ export class OIRecognizer
     }
     this.contextlessGestureDeferred.set(stroke.id, new DeferredPromise<TOIMessageEventContextlessGesture>())
     const pixelTomm = 25.4 / 96
+    await this.initialized?.promise
     await this.send({
       type: "contextlessGesture",
       scaleX: pixelTomm,
@@ -640,6 +649,7 @@ export class OIRecognizer
     const message: TOIMessageEvent = {
       type: "waitForIdle",
     }
+    await this.initialized?.promise
     await this.send(message)
     return this.waitForIdleDeferred?.promise
   }
@@ -691,6 +701,7 @@ export class OIRecognizer
       type: "undo",
       changes
     }
+    await this.initialized?.promise
     await this.send(message)
     return this.undoDeferred?.promise
   }
@@ -707,6 +718,7 @@ export class OIRecognizer
       type: "redo",
       changes
     }
+    await this.initialized?.promise
     await this.send(message)
     return this.redoDeferred?.promise
   }
@@ -725,6 +737,7 @@ export class OIRecognizer
       partId: this.currentPartId,
       mimeTypes
     }
+    await this.initialized?.promise
     await this.send(message)
     const exports = await Promise.all(mimeTypes.map(mt => this.exportDeferredMap.get(mt)!.promise))
     return Object.assign({}, ...exports)
@@ -733,6 +746,7 @@ export class OIRecognizer
   async clear(): Promise<void>
   {
     this.clearDeferred = new DeferredPromise<void>()
+    await this.initialized?.promise
     await this.send({
       type: "clear"
     })
